moved queries to general designer
