+++ Localize/lang/strings.json
@@ -165,7 +165,6 @@
   "2OQU3/": "Cannot run after both trigger and action",
   "2On4Xu": "Code view tab",
   "2TMGk7": "Managed identity",
-  "2XH9oW": "Back",
   "2ZfzaY": "Select existing",
   "2adqQ4": "Maximum interval",
   "2cVEMV": "Returns a single value matching the key name from form-data or form-encoded action output",
@@ -204,7 +203,6 @@
   "3QXY3z": "Replacing an existing schema with an incompatible schema might create errors in your map.",
   "3RoD4h": "Returns the collection in reverse order",
   "3ST5oT": "You're creating an accelerator template!",
-  "3Wcqsy": "Next",
   "3X4FHS": "Choose the type of user input",
   "3Xf/4S": "Swagger endpoint",
   "3Y8a6G": "Required parameters {parameters} not set or invalid",
@@ -1555,7 +1553,6 @@
   "_2OQU3/.comment": "error message for running after both trigger and action",
   "_2On4Xu.comment": "An accessibility label that describes the code view tab",
   "_2TMGk7.comment": "Managed Identity Label",
-  "_2XH9oW.comment": "Back button",
   "_2ZfzaY.comment": "Select existing option",
   "_2adqQ4.comment": "title for retry maximum interval setting",
   "_2cVEMV.comment": "Label for description of custom formDataValue Function",
@@ -1594,7 +1591,6 @@
   "_3QXY3z.comment": "Message bar warning about replacing existing schema",
   "_3RoD4h.comment": "Label for description of custom reverse Function",
   "_3ST5oT.comment": "Title for the toaster after adding workflows.",
-  "_3Wcqsy.comment": "Next button",
   "_3X4FHS.comment": "Button to choose data type of the dynamically added parameter",
   "_3Xf/4S.comment": "Swagger endpoint input label",
   "_3Y8a6G.comment": "Error message to show when required parameters are not set or invalid",
@@ -3251,7 +3247,6 @@
   "_mgD2ZT.comment": "The tab label for the monitoring parameters tab on the operation panel",
   "_mjS/k1.comment": "description of suppress woers setting",
   "_mlU+AC.comment": "Header for the connections panel",
-  "_mlxD6R.comment": "Finish  button",
   "_mnuwWm.comment": "Warning body for when unable to parse schema",
   "_mpFlLc.comment": "Mainframe Modernization category",
   "_mqVL/E.comment": "The tab label for the add actions tab on the connector panel",
@@ -3362,7 +3357,6 @@
   "_pH2uak.comment": "Label to collapse",
   "_pH6ubt.comment": "Column header for accessing connection-related details",
   "_pJJ3x8.comment": "Seach source or target nodes",
-  "_pK0Ir8.comment": "Export with warnings button",
   "_pOTcUO.comment": "Required object parameter to be converted to array using createArray function",
   "_pOVDll.comment": "Error validation message for Integers",
   "_pRrzwt.comment": "Error shown when the author (By) field is missing",
@@ -3643,7 +3637,6 @@
   "_wPi8wS.comment": "Accessibility label indicating that the value is not set",
   "_wPjnM9.comment": "Text for button to paste a parallel action from clipboard",
   "_wPlTDB.comment": "Full path of current node",
-  "_wPzyvX.comment": "Export button",
   "_wQcEXt.comment": "Required parameters for the custom Replace Function",
   "_wQsEwc.comment": "Required length parameter to obtain substring",
   "_wT/gMB.comment": "Description for featured connectors field",
@@ -4248,7 +4241,6 @@
   "mgD2ZT": "Summary",
   "mjS/k1": "Limit Logic Apps to not include workflow metadata headers in the outgoing request.",
   "mlU+AC": "Connections",
-  "mlxD6R": "finish",
   "mnuwWm": "This error might mean that the agent parameter schema is incorrectly set up.",
   "mpFlLc": "Mainframe Modernization",
   "mqVL/E": "Select actions",
@@ -4359,7 +4351,6 @@
   "pH2uak": "Collapse",
   "pH6ubt": "Details",
   "pJJ3x8": "Search nodes",
-  "pK0Ir8": "Export with warnings",
   "pOTcUO": "Required. The values to combine into an array.",
   "pOVDll": "Enter a valid integer.",
   "pRrzwt": "By field is required for publish.",
@@ -4640,7 +4631,6 @@
   "wPi8wS": "----",
   "wPjnM9": "Paste a parallel action",
   "wPlTDB": "Full path",
-  "wPzyvX": "Export",
   "wQcEXt": "Required. The string that is searched for parameter 2 and updated with parameter 3, when parameter 2 is found in parameter 1.",
   "wQsEwc": "Required. The length of the substring.",
   "wT/gMB": "Key services this template integrates with.",
