+++ libs/designer/src/lib/core/templates/utils/queries.ts
@@ -1,8 +1,7 @@
 import type { UseQueryResult } from '@tanstack/react-query';
 import { useQuery } from '@tanstack/react-query';
 import { getConnector, getOperation } from '../../queries/operation';
-import type { LogicAppResource, Resource } from '@microsoft/logic-apps-shared';
-import { ResourceService, TemplateService } from '@microsoft/logic-apps-shared';
+import { TemplateService } from '@microsoft/logic-apps-shared';
 import { getReactQueryClient } from '../../ReactQueryProvider';
 
 export interface ConnectorInfo {
@@ -49,73 +48,6 @@ export const useConnectorInfo = (
       refetchOnMount: false,
       refetchOnWindowFocus: false,
       refetchOnReconnect: false,
-    }
-  );
-};
-
-export const useSubscriptions = (): UseQueryResult<Resource[], unknown> => {
-  return useQuery(
-    ['templateQueries', 'subscriptions'],
-    async () => {
-      return ResourceService().listSubscriptions();
-    },
-    {
-      cacheTime: 1000 * 60 * 60 * 24,
-      refetchOnMount: false,
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-    }
-  );
-};
-
-export const useResourceGroups = (subscriptionId: string): UseQueryResult<Resource[], unknown> => {
-  return useQuery(
-    ['templateQueries', 'resourcegroups', subscriptionId],
-    async () => {
-      return ResourceService().listResourceGroups(subscriptionId);
-    },
-    {
-      cacheTime: 1000 * 60 * 60 * 24,
-      refetchOnMount: false,
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-      enabled: !!subscriptionId,
-    }
-  );
-};
-
-export const useLocations = (subscriptionId: string): UseQueryResult<Resource[], unknown> => {
-  return useQuery(
-    ['templateQueries', 'locations', subscriptionId],
-    async () => {
-      return ResourceService().listLocations(subscriptionId);
-    },
-    {
-      cacheTime: 1000 * 60 * 60 * 24,
-      refetchOnMount: false,
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-      enabled: !!subscriptionId,
-    }
-  );
-};
-
-export const useLogicApps = (
-  subscriptionId: string,
-  resourceGroup: string,
-  enabled: boolean
-): UseQueryResult<LogicAppResource[], unknown> => {
-  return useQuery(
-    ['templateQueries', 'logicapps', subscriptionId, resourceGroup],
-    async () => {
-      return ResourceService().listLogicApps(subscriptionId, resourceGroup);
-    },
-    {
-      cacheTime: 1000 * 60 * 60 * 24,
-      refetchOnMount: false,
-      refetchOnWindowFocus: false,
-      refetchOnReconnect: false,
-      enabled: enabled && !!subscriptionId && !!resourceGroup,
     }
   );
 };
